Refactor: rename engine files for consistency (#640)
Renames engine module files to shorter, clearer names:
- parameter_optimizer.py -> parameters.py
- slot_builder.py -> slots.py
- decision_outcome_tracker.py -> outcomes.py
- slot_schedule.py unchanged

Also renames corresponding test files to match.

Updates all import statements across codebase.
Storage keys preserved for backward compatibility.
