+++ custom_components/localshift/engine/__init__.py
@@ -9,11 +9,6 @@ from ..forecast.solar import (
     get_solar_for_15min_slot,
     sum_solar_before_target,
 )
-from .decision_outcome_tracker import (
-    DecisionOutcomeTracker,
-    DecisionRecord,
-    PerformanceMetrics,
-)
 from .excess_solar import ExcessSolarEngine
 from .excess_solar_signals import ExcessSolarSignalsEngine
 from .optimization_controller import (
@@ -22,7 +17,12 @@ from .optimization_controller import (
     OptimizationController,
 )
 from .optimizer_facade import OptimizerFacade
-from .parameter_optimizer import ParameterOptimizer
+from .outcomes import (
+    DecisionOutcomeTracker,
+    DecisionRecord,
+    PerformanceMetrics,
+)
+from .parameters import ParameterOptimizer
 from .pattern_analyzer import (
     BiasCorrection,
     PatternAnalyzer,
@@ -35,7 +35,7 @@ from .price_calculator import (
     get_price_for_slot_or_none,
 )
 from .price_signal_engine import PriceSignalEngine
-from .slot_builder import SlotBuilder, SlotBuildMetadata
+from .slots import SlotBuilder, SlotBuildMetadata
 from .soc_simulator import SocSimulator
 from .spike_analyzer import SpikeAnalyzer
 from .utils import (
