+++ custom_components/localshift/engine/outcomes.py
