+++ custom_components/localshift/engine/slots.py
