+++ tests/learning/test_learning_integration.py
@@ -9,14 +9,14 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from custom_components.localshift.engine.decision_outcome_tracker import (
+from custom_components.localshift.engine.outcomes import (
     DecisionOutcomeTracker,
     DecisionRecord,
 )
 from custom_components.localshift.engine.optimization_controller import (
     OptimizationController,
 )
-from custom_components.localshift.engine.parameter_optimizer import (
+from custom_components.localshift.engine.parameters import (
     ParameterOptimizer,
 )
 from custom_components.localshift.engine.pattern_analyzer import (
