+++ tests/test_parameters.py
@@ -5,10 +5,10 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from custom_components.localshift.engine.decision_outcome_tracker import (
+from custom_components.localshift.engine.outcomes import (
     DecisionRecord,
 )
-from custom_components.localshift.engine.parameter_optimizer import (
+from custom_components.localshift.engine.parameters import (
     ParameterOptimizer,
 )
 from custom_components.localshift.const import (
